update-parameters.py: fix the cache directory
Use $XDG_CACHE_HOME instead of $MC_ACCOUNT_DIR to find the file
mcp-test-diverted-account-plugin.conf because the test-only plugin
mcp-account-diversion uses g_get_user_cache_dir().

This is a no-op because both variables have the same value, as defined in
tests/twisted/Makefile.am:

BASIC_TESTS_ENVIRONMENT = \
	MC_ACCOUNT_DIR=@abs_top_builddir@/tests/twisted/tmp-$(TMPSUFFIX) \
	XDG_CACHE_HOME=@abs_top_builddir@/tests/twisted/tmp-$(TMPSUFFIX) \

However, this was not semantically correct, and the two directories could be
different in the future, with installed tests for example.