Account: only set ChangingPresence if we actually are
Previously, ChangingPresence could become True on accounts which were
not Enabled, which makes no sense: we're not changing our presence in
that case. It happened if you requested an offline presence on a
disabled account whose current RequestedPresence is something online; or
if you disabled an Enabled account whose RequestedPresence (and
CurrentPresence) was offline.

Here we change the implementation of setting RequestedPresence to only
set ChangingPresence to True if it actually takes some action. There are
situations where the presence has "changed" (maybe you changed the
message associated with being offline) but no change actually needs to
be applied.