Add a GSettings schema with use-conn key
When the NM/ConnMan code lived in Empathy, there was a GSettings key
which could be flipped to false to make Empathy always assume it had a
network connection.

This schema provides one GSettings key,
im.telepathy.MissionControl.FromEmpathy use-conn, which happens to use
the same dconf path as the Empathy key uses. This should make the two
GSettings keys aliases for the same underlying value in dconf; we get a
migration path for free.