Bind use-conn GSettings key to ConnectivityMonitor
This GSettings key does the same thing as the Empathy one it's an alias
for does: if use-conn is False, the network state from NM/ConnMan is
ignored, and MC will allow you to try to sign in even if there is
supposedly no network connection! It's pretty great.

(When I first committed this, it didn't work. Now it does. I don't know
why.)