Account: don't crash on Get('Parameters') if _dup_parameters fails
If the connection manager is not installed, or is installed but doesn't
have the protocol for the account, then _mcd_account_dup_parameters()
returns %NULL. (As the comment I've added says, I think this is
ridiculous—it's fall-out from MC not storing the types of parameters.)

As a result, Get('Parameters') or GetAll() would crash, because
dbus-glib would try to serialize a NULL GHashTable and blow up. (See
<https://bugs.freedesktop.org/show_bug.cgi?id=44939>.) So with this
patch, we handle NULL as "no parameters" for the sake of D-Bus
properties. (Because this situation should only arise if the account is
invalid, we also warn if _dup_parameters() returns NULL for valid
accounts.)

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>