_mcd_client_proxy_add_cap_tokens: switch from "add" to "set" semantics
In both the cases where it's called (when we get a reply to GetAll,
and when we've parsed a .client file), we actually want to take the
new set, not union it with an existing set.

Also, this will be easier to deal with when we stop (ab)using handles
for this.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=49753