dispatcher: set channel state back to DISPATCHED after delegating & presenting
_mcd_client_proxy_handle_channels sets the McdChannel state to
HANDLER_INVOKED, but when we were calling this in both delegate and
present operations, the state was not being set back to the DISPATCHED
state afterwards.

The state being wrong meant that dispatching a channel after if had
been delegated or presented would not happen, as MC thought we were
still waiting for a handler response.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>