Migrate from Butterfly to Haze even if Butterfly is no longer installed
Mission Control has a bug/design flaw/whatever in which it doesn't
know the types of parameters without looking at the corresponding
connection manager, which doesn't work if that connection manager has
been uninstalled. This turns out to be bad for the upgrade from
Debian 6 to Debian 7, which upgrades MC to a version that will
perform the migration from Butterfly to Haze, but also uninstalls
Butterfly...

Since we know what the type of 'account' is, with a bit of refactoring
we can tell McdAccount that we know it's a string, get its value
anyway, and get on with our lives.

Bug-Debian: http://bugs.debian.org/686835
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Xavier Claessens <xavier.claessens@collabora.co.uk>