mcd_storage_dup_value: don't return non-NULL with error set
The practical effect of this is:

* addressing_set_uri_scheme_association() needs modification to cope
  with a NULL return, which previously never happened (an empty
  string-list was returned on error)

* addressing_get_uri_schemes() needs the same change

* mcd_account_get_parameter() treats parameters whose value in the
  keyfile cannot be interpreted as the desired type as if they were
  unset; previously, it treated them as 0, False, "", [], etc.

The old behaviour of mcd_account_get_parameter() seems undesirable:
we shouldn't just make up parameter values. If the parameter is boolean
"require-tls" and its value gets set to 2 for some reason, we should
invalidate the account, rather than setting it to False and carrying on!