McdAccount: store automatic presence as a single attribute
This will be necessary to make its change-notification work correctly.

Use tests/twisted/account-manager/auto-connect.py to exercise the
automatic migration.

account-storage/default-keyring-storage.py is broken by this change,
because it asserts that keyfiles from a lower-priority XDG location
are not automatically migrated to a higher-priority location until
something actually changes, and the migration of AutomaticPresenceType
is a change. To avoid that, put the automatic presence in lower-priority
keyfiles in that test in the new format already.