enable_fakecm_account: copy parameters and alter the copy
Mutable defaults for arguments are dangerous. If we call
enable_fakecm_account twice with default arguments, when we alter
expect_before_connect/expect_after_connect, we are actually altering
the default! Python doesn't construct a new empty list for us every
time. The C equivalent would be a static GPtrArray * or something.