McdConnection: build up our sets of emergency numbers over time
To facilitate this, use nicer data structures.

Most importantly, this means we don't critical whenever we get more
than one distinct service point, as Jonny noticed while porting to
Telepathy 1.0. It might slightly defeat the intention of the
ServicePointsChanged signal, but in practice the list is probably
never going to shrink, and if it does, it's probably wise to
keep considering its old members to be an emergency number (the effect
that that has is that plugins aren't allowed to influence channel
requests).

Also, don't leak the sets of emergency numbers when finalized.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=55773
Reviewed-by: Marco Barisione <marco.barisione@collabora.co.uk>