Only ignore passwords in gnome-keyring if Empathy has copied them
When passwords are stored in the Parameters by MC and no
SASLAuthentication handler is used, e.g. under Sugar, we currently
save passwords into gnome-keyring, if enabled at compile time; so, we
want to read them back out of gnome-keyring.

This conflicts slightly with wanting to ignore passwords in gnome-keyring
that are left over from an Empathy 3.0 schema migration that failed
due to fd.o #42088. To avoid failure to authenticate/re-prompt if the
password has been changed on the service side, ignore passwords that
are going to be deleted; but to avoid breaking Sugar, don't ignore
any other passwords we might find in gnome-keyring.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=59468