tests: avoid an out-of-range int literal
In this particular test case, the expected result according to the flags
is "it fails", so the second column (expected converted value) is not
relevant.

This apparently causes build warnings (fatal by default) on ILP32
platforms, because 2147483648 is out-of-range for a signed int or
signed long on such platforms. It's certainly also out-of-range for
the struct member it's initializing, which is a gint32.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=64083