service points: do not treat their handles as special
Requesting channels by TargetHandle seems rather perverse when we've
gone to some length to ensure that identifiers are always at least
as available, and in particular, if the user of a Telepathy-based
phone or something types 911 into a keypad, the request is going
to be TargetID-based. telepathy-spec 1.0 should specify this.

https://bugs.freedesktop.org/show_bug.cgi?id=69176
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>