connectivity test: make sure we've caught up on GDBus messages
Otherwise, the message from the FakeNetworkMonitor saying we've gone
offline might not be received until after the message from the fake NM
saying we've gone online - which would be a problem. There's no
side-effect we can wait for here, so we just have to make sure the
GDBus queue has been processed.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=69585
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>