avatar-refresh test: subsume avatar-persist, and test more situations
We have some sort of combinatorial explosion going on here, and it
seems best to test it in a somewhat systematic way:

* is the protocol one where avatars persist on the server (Gabble)
  or not (Salut)?
* if it's like Gabble, does it download our own avatar token
  before signalling CONNECTED (as I suspect Haze does), or
  on-demand after GetKnownAvatarTokens (as Gabble appears to)?
* if it's like Gabble, is the server storing an avatar for us?
* in either case, do we have an avatar stored locally, and has
  it previously been uploaded or not?

In addition, the avatar-refresh and avatar-persist tests exercised
migration from ~/.missioncontrol and a low-priority XDG_DATA_DIRS entry
(respectively) to ~/.local/share. I didn't do that in a loop, because
it isn't applicable in all cases and would lead to even more
combinations - testing each case once should be enough.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=69885
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>