McdAccount: go back to using GetKnownAvatarTokens
Sadly, contact attributes aren't enough to distinguish between
"this protocol doesn't store avatars and you haven't re-sent your
avatar since you last connected", "this protocol stores avatars but
the CM hasn't checked for your current avatar yet", and "this protocol
stores avatars, but there is no avatar on the server for you".

GetKnownAvatarTokens specifically excludes the middle option (blocking
on a server round-trip if it needs to), and uses "avatar token missing"
for the first and "avatar token empty" for the last.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=69885
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>