McdAccount: go back to using GetKnownAvatarTokens
Sadly, contact attributes aren't enough to distinguish between
"this protocol doesn't store avatars and you haven't re-sent your
avatar since you last connected", "this protocol stores avatars but
the CM hasn't checked for your current avatar yet", and "this protocol
stores avatars, but there is no avatar on the server for you".

GetKnownAvatarTokens specifically excludes the middle option (blocking
on a server round-trip if it needs to), and uses "avatar token missing"
for the first and "avatar token empty" for the last.

Tested to work with the telepathy-glib 0.20 branch, and with the 0.22
branch when fdo#70010 is fixed.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=69885
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>
[This is a cherry-pick of commit 4f4ed24 from master, omitting the
regression test change, to make sure the existing tests still pass -smcv]