already-has-channel test: alter how we delay preparation
We used to delay GetInterfaces and Get(CONNECTION, Interfaces), but in
recent telepathy-glib, we just use GetAll(CONNECTION), so this was
not effective. However, this was masked by the fact that the
ServicePoints code redundantly calls GetInterfaces. :-(

GetAll is also what we use to replace GetStatus, so connection won't
proceed if we delay that. However, we can also set up the situation
we're trying to test by making GetAll(CONNECTION) return promptly,
and delaying GetAll(REQUESTS) instead.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=54879
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>