Remove support for UPower
As of git master, it no longer has the API we use. On Linux, the
recommended way to be notified about suspend this week appears to
be logind, for which support was added in 5.15.1. Elsewhere, this
is a regression, but if UPower is deleting this functionality
we don't have much choice.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=70458
Reviewed-by: Colin Walters <walters@verbum.org>
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>