Replace McdAccount::always-on with the existing TpStorageRestrictionFlags
I used CANNOT_SET_PRESENCE to control access to ConnectAutomatically
as well as the obvious AutomaticPresence and RequestedPresence, because
RequestedPresence is not under the storage plugin's control - it is
derived from ConnectAutomatically and AutomaticPresence at runtime.

Use MCD_DBUS_PROP_SET_FLAG_ALREADY_IN_STORAGE as a way to bypass the
storage restriction flags, so that storage plugins themselves can
alter enabledness etc. even if they don't allow MC to change it.

The regression test for this initially failed, because toggled_cb()
did not pass that flag to the McdAccount: fix that too.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=71390