McdStorage: adjust IdentifyAccount error behaviour
We were ignoring failures if they were NotImplemented or ServiceUnknown,
but thinking about it more, we should probably ignore "most" errors
here: the only errors that should abort the account-creation attempt
are those that indicate that the intended parameters are unusable,
namely InvalidArgument and InvalidHandle (in its secondary role as
"invalid identifier-that-corresponds-to-a-handle").

Also add more debug messages here.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=27727