Make account deletion async (sort of), and imply a selective commit
This means we don't need to commit separately after each deletion,
and means account plugins don't have to have the concept of flagging
an account for "delete this later" - much rejoicing.

It also has the incidental benefit that we no longer use the C++
reserved word 'delete' in a header file.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=27727