dbus-account-plugin: fail tests on various invalid method calls
Methods that act on a single account should always be called on an
account that exists (if it doesn't, where did you get its name from?),
so treat it as an error if they are called with a nonexistent account.

commit() has a dual role here: the method that takes a non-NULL account
(which acts on a single account, as described), and the method that
takes a NULL account name (which is OK to call at any time, to commit
all of our 0-or-more accounts).

While we are not active, we don't claim to have any accounts at all,
so treat it as an error if any method is called with a non-NULL account
name while inactive.

I've temporarily made an exception for delete_async(), although that
exception is removed in a later patch on this branch. It's less bad
if delete_async() is called on an account that doesn't exist, because
if you do, the desired state ("account X doesn't exist") has already
been reached.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=27727