McdAccountManager: separate setup lock for AM from lock for new accounts
McdLoadAccountsData was being used for two separate things: either
the initial setup process for the AccountManager (not specific to an
account, blocking taking the AccountManager name) or the setup process
for an account being added by a storage plugin later on (specific to
an account, and hopefully after the AccountManager name has been taken).

Since release_load_accounts_lock() takes the AccountManager name in an
idempotent way, this happened to work most of the time. However,
if a storage plugin signals the addition of an account not in the
initial batch, and adding that account finishes before the setup
process for the initial batch, we'd take the name too early, and have
our D-Bus API before we should - that'd be bad.

Redo this so it makes more sense. We don't actually need a struct
for the "initial batch" case, because the McdAccountManager and maybe
a MigrateCtx is enough; just keep a count of locks in
McdAccountManagerPrivate.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=74185
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>