_mcd_account_dup_parameters: try to get parameters' types from backend
Now that I've deleted ExternalAccountStorage support, we have two
uses for this function:

* get the parameters to be passed to RequestConnection

* get the parameters for our own D-Bus API (PropertiesChanged,
  GetAll, etc.)

For the former, we should know the types already, because we should
already have a concrete CM/protocol in mind by the time we get here.

For the latter, ideally we shouldn't need the CM's types at all: if
the backend is storing parameters with types, it's arguably more
correct for Parameters to contain what the user stored, even if that
isn't an exact match for what the CM wants.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=71093
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>