mcd-account: unregister the account once we're done deleting it
The McdAccount object may still be alive for a short while because of pending
async operations. We want to get rid of the D-Bus object as soon as possible
and so, for example, a new Account object replacing the deleted one can be
created.

Fix https://bugs.freedesktop.org/show_bug.cgi?id=79827