tests: fix race condition in account-manager/avatar.py test
It seems that the order of the dbus signals is unpredictable,
so requiring it in the test sometimes makes it fail. It also
seems that the other checks above in the same unit test have
the same logic of not expecting a certain order in signal emission,
so this is probably the right thing to do.